Refactor package organization
