+++ app/src/main/java/agency/five/tmdb/data/database/Entities.kt
@@ -3,7 +3,7 @@ package agency.five.tmdb.data.database
 
 import agency.five.tmdb.domain.common.Actor
 import agency.five.tmdb.domain.common.PersonFunction
-import agency.five.tmdb.ui.components.MovieItemViewState
+import agency.five.tmdb.presentation.ui.components.MovieItemViewState
 import androidx.room.*
