+++ app/src/main/java/agency/five/tmdb/data/database/MovieDao.kt
@@ -1,6 +1,6 @@
 package agency.five.tmdb.data.database
 
-import agency.five.tmdb.ui.components.MovieItemViewState
+import agency.five.tmdb.presentation.ui.components.MovieItemViewState
 import androidx.room.*
 import kotlinx.coroutines.flow.Flow
