+++ app/src/main/java/agency/five/tmdb/data/repository/MovieRepositoryImpl.kt
@@ -7,7 +7,7 @@ import agency.five.tmdb.data.database.MoviePersonsFunctions
 import agency.five.tmdb.domain.common.Actor
 import agency.five.tmdb.domain.common.PersonFunction
 import agency.five.tmdb.domain.repointerfaces.MovieRepository
-import agency.five.tmdb.ui.components.MovieItemViewState
+import agency.five.tmdb.presentation.ui.components.MovieItemViewState
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.*
