+++ app/src/main/java/agency/five/tmdb/domain/repointerfaces/MovieRepository.kt
@@ -3,7 +3,7 @@ package agency.five.tmdb.domain.repointerfaces
 
 import agency.five.tmdb.domain.common.Actor
 import agency.five.tmdb.domain.common.PersonFunction
-import agency.five.tmdb.ui.components.MovieItemViewState
+import agency.five.tmdb.presentation.ui.components.MovieItemViewState
 import kotlinx.coroutines.flow.Flow
 
 interface MovieRepository {
