+++ app/src/main/java/agency/five/tmdb/presentation/navigation/Screens.kt
@@ -1,4 +1,4 @@
-package agency.five.tmdb.navigation
+package agency.five.tmdb.presentation.navigation
 
 sealed class Screens(val route: String) {
     object MainScreen : Screens("main_screen")  //add favorites screen and home screen
