+++ app/src/main/java/agency/five/tmdb/presentation/ui/components/ActorCards.kt
@@ -1,8 +1,8 @@
-package agency.five.tmdb.ui.components
+package agency.five.tmdb.presentation.ui.components
 
 import agency.five.tmdb.R
 import agency.five.tmdb.domain.common.Actor
-import agency.five.tmdb.ui.theme.Typography
+import agency.five.tmdb.presentation.ui.theme.Typography
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
