+++ app/src/main/java/agency/five/tmdb/presentation/ui/components/SearchBar.kt
@@ -1,7 +1,7 @@
-package agency.five.tmdb.ui.components
+package agency.five.tmdb.presentation.ui.components
 
 import agency.five.tmdb.R
-import agency.five.tmdb.ui.theme.Typography
+import agency.five.tmdb.presentation.ui.theme.Typography
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
