+++ app/src/main/java/agency/five/tmdb/presentation/ui/screens/DetailsScreen.kt
@@ -1,14 +1,14 @@
-package agency.five.tmdb.ui.screens
+package agency.five.tmdb.presentation.ui.screens
 
 
 import agency.five.tmdb.R
 import agency.five.tmdb.domain.common.Actor
 import agency.five.tmdb.domain.common.PersonFunction
-import agency.five.tmdb.ui.components.ActorCardsList
-import agency.five.tmdb.ui.components.MovieDetails
-import agency.five.tmdb.ui.components.MovieItemViewState
-import agency.five.tmdb.ui.theme.Typography
-import agency.five.tmdb.viewmodel.DetailsScreenViewModel
+import agency.five.tmdb.presentation.ui.components.ActorCardsList
+import agency.five.tmdb.presentation.ui.components.MovieDetails
+import agency.five.tmdb.presentation.ui.components.MovieItemViewState
+import agency.five.tmdb.presentation.ui.theme.Typography
+import agency.five.tmdb.presentation.viewmodel.DetailsScreenViewModel
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
