+++ app/src/main/java/agency/five/tmdb/presentation/ui/screens/MainScreen.kt
@@ -1,8 +1,8 @@
-package agency.five.tmdb.ui.screens
+package agency.five.tmdb.presentation.ui.screens
 
 import agency.five.tmdb.R
-import agency.five.tmdb.ui.theme.TmdbTheme
-import agency.five.tmdb.ui.theme.Typography
+import agency.five.tmdb.presentation.ui.theme.TmdbTheme
+import agency.five.tmdb.presentation.ui.theme.Typography
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
