+++ app/src/main/java/agency/five/tmdb/presentation/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package agency.five.tmdb.ui.theme
+package agency.five.tmdb.presentation.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
