+++ app/src/main/java/agency/five/tmdb/presentation/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package agency.five.tmdb.ui.theme
+package agency.five.tmdb.presentation.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
