+++ app/src/main/java/agency/five/tmdb/presentation/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package agency.five.tmdb.ui.theme
+package agency.five.tmdb.presentation.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.graphics.Color
